tests/Makefile: allow setting DIO_TEST_DEV in the environment
This change has two effects: 1. no error message is printed if
the directio_test_dev file doesn't exist, and 2. DIO_TEST_DEV
can be set in the environment, or on the "make" command line.
It's still possible to set DIO_TEST_DEV via the directio_test_dev
file if preferred.

Signed-off-by: Martin Wilck <mwilck@suse.com>