tests/Makefile: add test_clean target
This is a convenience target to remove only the test output files.
It's useful for those tests which require root permissions to run,
as "make" doesn't delete the .out file if a test fails (we could
force "make" to do so, but we don't want to do that, because usually
we want to inspect the output file when a test has failed). With
test_clean, we can run "make" as normal user, then "make test_clean"
and finally run "make" as root again.

Signed-off-by: Martin Wilck <mwilck@suse.com>