multipathd: uev_trigger(): handle incomplete ADD events
Udev may be killed after handling the ADD event for a multipath map,
but before handling the subsequent CHANGE event that populates the
udev data base with the device properties (e.g. during initrd processing).
If this happens, the ADD uevent sent during coldplug will only provide a
subset of the device properties. We need another CHANGE event to make the map
available to the system.

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>