multipath-tools: fix compilation errors on 32-bit musl
gcc on alpine Linux/i386 throws errors because the "tv_sec" element
of struct timespec is a time_t, which is a "long long" in that
environment. In general, time_t is signed. As we only use CLOCK_MONOTONIC,
which starts at boot time, a cast to long should be no problem, even
in 32bit environments.

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>