github workflows: add containerized / multi-arch tests
Until now, our CI only tested builds on Ubuntu, the default github
runner. This commit adds containerized build/test workflows, which can cover a
much larger range of distributions and environments.

I've distinguished "native" runs, where the architecture is compatible with
the architecture of the runners (assumed to be x86_64), and "foreign" runs
with different architectures. The former can run "make test" just like in
any ordinary environment. To run tests in "foreign arch" environments,
I use qemu-user for runtime emulation. That's done with the help of the
"multiarch/qemu-user-static" container. Runtime containers for testing need
to have the qemu-user-static binary (compiled for x86_64) for their target
architecture built in. In theory, we could do the same thing for build
containers too, but it would be grossly inefficient. Instead, I've focused
on Debian environments for the foreign architectures, relying on Debian's
nice multiarch / cross-compilation features. Compilation is run in a dedicated
cross-build container, and only the test runs are carried out in the emulation.

The set of tests currently includes:

native:
 - Debian Jessie, x86_64/i386 (gcc 4.9, clang 3.5, glibc 2.19)
 - Debian Buster, x86_64/i386 (gcc 8.3, clang 7.0, glibc 2.28)
 - Debian Sid, x86_64/i386 (gcc 10.2, clang 11.0, glibc 2.31)
 - Alpine, x86_64/i386 (gcc 10.2, clang 10.0, musl libc 1.2)
 - Fedora 34, x86_64 (gcc 11.0, clang 12.0, glibc 2.33)

foreign:
 - Debian Buster, ppc64le/aarch64/s390x

This covers a rather broad range of compiler and C library versions and
should be fine for some time to come.

Note: In theory, it would be possible to just fetch base containers
via github actions, and install the dependencies as part of the build
procedure. But that would be quite resource-intensive and slow. Therefore
I've decided to use pre-built containers. The current container setup
fetches containers from my docker hub repository. The containers there
(multipath-build-$os-$arch and multipath-run-$os-$arch) come from my
"build-multipath" repository (https://github.com/mwilck/build-multipath),
and are created via github actions, too. The upload of the built container
images to docker hub requires the use of tokens and secrets.

I'd considered adding these container definitions and workflows to the
multipath-tools repository. We'd just need to create reasonable rules
for running the respective workflows. I expect these container images
to remain relatively stable; it makes no sense to rebuild the images
for every multipath-tools commit.

Tell me if you want this in the multipath-tools repo, and if you're ok
with hosting the images in my docker hub repo.

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>