libmultipath: add optional wakeup functionality to lock.c
Have struct mutex_lock take an optional wakeup function.
unlock() is renamed to __unlock() in order to prevent it from
being called by mistake.

This changes offsets in "struct vectors", requiring a major
libmultipath version bump. While the strucure is already changed,
in order to avoid this in the future, move the lock to the end
of "struct vectors".

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>