multipathd: uxlsnr: add timeout handling
Our ppoll() call needs to wake up when a client request times out.
This logic can be added by determining the first client that's about
to time out. The logic in handle_client() will then cause a timeout
reply to be sent to the client. This is more client-friendly
as the client timing out without receiving a reply.

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>