Add a -1 flag for disabling systemd integration
Support disabling systemd notification integration at run time.

In some deployments of mulitpathd, systemd is ultimately used but not
in stage-1 of boot. In this scenario multipathd is used to mount the
root volume, and then switch_root runs, and we call systemd.

For this use case it is possible to "just" have two copies of
multipathd around, one with and one without systemd support, but it
is a bit tricky and annoying to juggle the two.