multipath-tools: support ABI testing with libabigail
Use abidw and abidiff (https://sourceware.org/libabigail/) to
generate a formal representation of our ABI, and check for changes.
This will reduce the amount of attention required to detect and
track library version changes.

To check for differences, run "make abi" on some branch, rename the
"abi" directory to "reference-abi", checkout a different branch,
and run "make abi-test".

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>