multipath-tools: add github workflow to save and check ABI
This adds a workflow that saves the ABI of libmultipath and the
other libraries, and optionally tests it against a known-good state,
which is taken from the configurable ABI_BRANCH. If the ABI differs,
the workflow fails, and the abidiff output is saved in GH actions
as artifact "abi-test".

To configure the reference branch, set the repository secret ABI_BRANCH to the
name of the branch that contains the ABI reference. The default is "master".

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>