libmultipath: introduce width argument for pretty-printing functions
Instead of modifying the global variables to store the field width for
pretty-printed table-formatted output, use field-width arrays passed
down from the caller. The calling function needs to allocate the array
using alloc_path_layout() or alloc_multipath_layout(), and initialize the
layout with the data to be printed. This is no different than before,
except that the field width is now stored in local variables of the caller.

The type used for column width, fieldwidth_t, is defined to unsigned char.
This limits the size of fields (padding in table columns) to 255 characters.
That's plenty even for lengthy WWID formats (as WWID_SIZE is 128). The typedef
could be changed in the future if necessary.

In some functions, namely _print_multipath_topology() and
print_foreign_topology(), field widths need to be initialized. This was true
before this patch already, but it wasn't done as the necessity wasn't
apparent. Essentially we were using pre-calculated field widths in these
functions, or just 0 if no field width calculation had been done previously.

Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>