multipathd.service: remove LimitCORE=infinity directive
This would normally be set by general system policy. According
to systemd-system.conf(5), it is set to infinity by default
anyway in systemd. If a user changes this in systemd-system.conf,
the setting should also apply to multipathd. Users who don't wont
this can still use drop-ins to change the specific settings.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>